backend/requirements-railway.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0

# HTTP client
aiohttp==3.9.1

# Image processing (optional for production)
Pillow>=10.0.0

# Remove pytesseract for now as it requires system dependencies
# pytesseract==0.3.10

# AI dependencies (optional - can be added later)
# google-generativeai==0.3.2